Fixed a few bugs in RDF/XML parser, in order of significance:
1. Parsing of multiple attributes on property nodes now fixed.
2. Reification based on rdf:ID only produces one reified statement.
3. Whitespacing in rdf:Literal fixed (no longer trimmed).
4. rdf:Literals work in rdf:li elements.
5. Duplicate rdf:IDs that yield different URIs not flagged as duplicates any more.

The parser passes all positive (non-entailment) tests for the offical W3C test-cases. However, it incorrectly reports a warning when non-ASCII characters are used in rdf:ID (a problem with XMLRegexes -- parsed content is fine and warning can be ignored).

The parser does not pass many of the negative tests: in other words, the parser may not fail cleanly if the RDF/XML document has invalid syntax.